cleaned up confetti
